// src/app/model/camp-menu.ts
export class CampMenu {
    date: string;
    breakfastFood?: number;
    breakfastDrink?: number;
    snackLowFood?: number;
    snackLowDrink?: number;
    lunchFood?: number;
    lunchDrink?: number;
    snackHighFood?: number;
    snackHighDrink?: number;
    dinnerFood?: number;
    dinnerDrink?: number;
}